app/requirements.txt
blinker==1.8.2
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==3.0.4
